requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
cloudscraper>=1.2.71
numpy>=1.25.0
scipy>=1.11.0
rich>=13.7.0
python-dotenv>=1.0.0
python-telegram-bot[job-queue]>=21.0
